backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.4.0
geopy==2.3.0
PyPDF2==3.0.1
elevenlabs==0.2.27
pydub==0.25.1
numpy==1.26.4
Pillow==10.2.0
requests==2.31.0
openai
python-dotenv


